% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{clean_and_prepare}
\alias{clean_and_prepare}
\title{Tidy up the power usage data by formatting the date correctly and saving month, weekday and year variables separately.}
\usage{
clean_and_prepare(data)
}
\arguments{
\item{data}{A power usage data set.}
}
\value{
A cleaned data frame with a few extra columns to make visualization easier.
}
\description{
Tidy up the power usage data by formatting the date correctly and saving month, weekday and year variables separately.
}
